LayoutTests/css3/scroll-snap/scroll-padding-overflow-paging-expected.txt

PASS paging moved container
PASS paging moved padded container
PASS container.scrollTop is within 1 of 140
PASS paging moved padded container
PASS container.scrollTop is within 1 of 140
PASS paging moved padded container
PASS container.scrollTop is within 1 of 132
PASS successfullyParsed is true

TEST COMPLETE
